Gallery Files
Renamed from MoNETFrontend to MoNETGallery. Page looks better, in line
with website. Arrows are my graphics. Currently working on creating
multiple pages of thumbnails so that images can be viewed on the same
page and the only thing that refreshes is the thumbs. Also to prevent
the page from being populated by tons of thumbnails, causing the user to
have to scroll. Goals, fix it, get the next/prev arrows to work. Some of
the .js files do not work and may be deleted later. Right now the images
are hardcoded with js, however I would like to implement the ability to
auto pull images and descriptions (need to add a html element for them)
from a DB of stored images from the robot, so we don't have to remember
to hardcode them ourselves. Then onto cross-browser and mobile
compatibility, may ditch gallery in favor of a thumb-slider for a
mobile-only experience. I fear, even bootstrap  or testresponsiveness.js
might not make it very small screen friendly.